Refactor out SubiquityServer paths
Server directory and Unix socket paths
