+++ packages/subiquity_client/lib/src/server.dart
@@ -1,17 +1,12 @@
 import 'dart:io';
 
 import 'package:meta/meta.dart';
-import 'package:package_config/package_config.dart';
 import 'package:path/path.dart' as p;
-import 'package:ubuntu_logger/ubuntu_logger.dart';
 import 'package:xdg_directories/xdg_directories.dart' as xdg;
 
 import 'endpoint.dart';
-
-enum ServerMode { LIVE, DRY_RUN }
-
-/// @internal
-final log = Logger('subiquity_server');
+import 'server/common.dart';
+import 'server/paths.dart';
 
 abstract class SubiquityServer {
   Process? _serverProcess;
@@ -36,45 +31,6 @@ abstract class SubiquityServer {
   // The name of the server's Python module.
   String get _pythonModule;
 
-  // The path of the socket is fixed in live-run mode. In dry-run mode, it needs
-  // to be resolved based on the location of the `subiquity_client` package.
-  Future<String> _getSocketPath(ServerMode mode) async {
-    if (mode == ServerMode.DRY_RUN) {
-      // Use a relative path to avoid hitting AF_UNIX path length limit because
-      // <path/to/ubuntu-desktop-installer>/packages/subiquity_client/subiquity/.subiquity/socket>
-      // grows easily to more than 108-1 characters (char sockaddr_un::sun_path[108]).
-      return p.relative(p.join(await _getSubiquityPath(), '.subiquity/socket'));
-    }
-    return '/run/subiquity/socket';
-  }
-
-  String? _subiquityPath;
-
-  // Returns the location of the local subiquity submodule.
-  Future<String> _getSubiquityPath() async {
-    return _subiquityPath ??= await findSubiquityPath();
-  }
-
-  // Finds local subiquity relative to the `subiquity_client` Dart package.
-  static Future<String> findSubiquityPath() async {
-    Object? error;
-    final config = await findPackageConfig(
-      Directory.current,
-      onError: (e) => error = e,
-    );
-    final package = config?.packages
-        .firstWhere((package) => package.name == 'subiquity_client');
-    if (package == null) {
-      log.warning(
-          'Unable to find the subiquity_client package. '
-          'Falling back to the current working dir: ${Directory.current.path}',
-          error);
-    } else {
-      log.debug('Found subiquity_client in ${package.root.path}');
-    }
-    return p.join(package?.root.path ?? Directory.current.path, 'subiquity');
-  }
-
   // Prefer local curtin and probert python modules that are pinned to the
   // correct versions
   Map<String, String> _pythonPath(String subiquityPath) {
@@ -87,7 +43,7 @@ abstract class SubiquityServer {
 
   Future<Endpoint> start(ServerMode serverMode,
       {List<String>? args, Map<String, String>? environment}) async {
-    final socketPath = await _getSocketPath(serverMode);
+    final socketPath = await SubiquityPaths.getSocketPath(serverMode);
     final endpoint = Endpoint.unix(socketPath);
     if (_shouldStart(serverMode)) {
       var subiquityCmd = <String>[
@@ -107,7 +63,7 @@ abstract class SubiquityServer {
 
   Future<void> _startSubiquity(ServerMode serverMode, List<String> subiquityCmd,
       Map<String, String>? environment) async {
-    final subiquityPath = await _getSubiquityPath();
+    final subiquityPath = await SubiquityPaths.getSubiquityPath();
     String? workingDirectory;
     // try using local subiquity
     if (Directory(subiquityPath).existsSync()) {
