+++ packages/subiquity_client/lib/src/server/common.dart
@@ -1,0 +1,6 @@
+import 'package:ubuntu_logger/ubuntu_logger.dart';
+
+enum ServerMode { LIVE, DRY_RUN }
+
+/// @internal
+final log = Logger('subiquity_server');
