+++ packages/subiquity_client/lib/subiquity_server.dart
@@ -1,4 +1,6 @@
 library subiquity_server;
 
 export 'src/endpoint.dart';
-export 'src/server.dart' hide log;
+export 'src/server/common.dart' hide log;
+export 'src/server/paths.dart';
+export 'src/server.dart';
